backend/requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
django==2.2.16
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.1
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.15
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itypes==1.2.0
jinja2==2.11.2
markupsafe==1.1.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.3
python-telegram-bot==12.8
python3-openid==3.2.0
pytz==2020.1
pyyaml==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
